@@ -1,17 +1,19 @@
 package world.bentobox.bank.commands.user;
 
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 import net.milkbowl.vault.economy.EconomyResponse;
 import net.milkbowl.vault.economy.EconomyResponse.ResponseType;
+import org.bukkit.Bukkit;
+import org.bukkit.entity.Player;
 import world.bentobox.bank.commands.AbstractBankCommand;
 import world.bentobox.bank.data.Money;
 import world.bentobox.bentobox.api.commands.CompositeCommand;
 import world.bentobox.bentobox.api.localization.TextVariables;
 import world.bentobox.bentobox.api.user.User;
+import world.bentobox.bentobox.database.objects.Island;
 import world.bentobox.bentobox.hooks.VaultHook;
+import world.bentobox.bentobox.managers.RanksManager;
 
 /**
  * @author tastybento
@@ -50,21 +52,27 @@ public class DepositCommand extends AbstractBankCommand {
         if (response.type == ResponseType.SUCCESS) {
             addon.getBankManager().deposit(user, value, getWorld()).thenAccept(result -> {
                 switch (result) {
-                case FAILURE_LOAD_ERROR:
-                    user.sendMessage("bank.errors.bank-error");
-                    break;
-                case FAILURE_LOW_BALANCE:
-                    user.sendMessage("bank.errors.low-balance");
-                    break;
-                case FAILURE_NO_ISLAND:
-                    user.sendMessage("general.errors.no-island");
-                    break;
-                case SUCCESS:
-                    user.sendMessage("bank.deposit.success", TextVariables.NUMBER, vault.format(addon.getBankManager().getBalance(island).getValue()));
-                    break;
-                default:
-                    break;
+                    case FAILURE_LOAD_ERROR -> user.sendMessage("bank.errors.bank-error");
+                    case FAILURE_LOW_BALANCE -> user.sendMessage("bank.errors.low-balance");
+                    case FAILURE_NO_ISLAND -> user.sendMessage("general.errors.no-island");
+                    case SUCCESS -> {
+                        user.sendMessage("bank.deposit.success", TextVariables.NUMBER, vault.format(addon.getBankManager().getBalance(island).getValue()));
 
+                        if(!addon.getSettings().isSendBankAlert()) return;
+
+                        Island island = addon.getIslands().getIsland(getWorld(), user);
+
+                        final Set<UUID> members = island.getMemberSet(RanksManager.MEMBER_RANK);
+                        for (UUID member : members) {
+                            final Player player = Bukkit.getPlayer(member);
+
+                            if (player == null || user.getUniqueId().equals(member)) continue;
+
+                            User otherUser = User.getInstance(player);
+
+                            otherUser.sendMessage("bank.deposit.alert", "[name]", user.getName(), "[number]", String.valueOf(value.getValue()));
+                        }
+                    }
                 }
             });
             return true;
@@ -77,6 +85,6 @@ public class DepositCommand extends AbstractBankCommand {
     public Optional<List<String>> tabComplete(User user, String alias, List<String> args) {
         VaultHook vault = addon.getVault();
         String balance = String.valueOf(vault.getBalance(user, getWorld()));
-        return Optional.of(Collections.singletonList(balance));
+        return Optional.of(List.of(balance, "all"));
     }
 }
